merging

+++ src/main/java/com/example/ngoprojectbackend/controller/EventMgmtController.java
@@ -1,5 +1,6 @@
 package com.example.ngoprojectbackend.controller;
 
+import com.example.ngoprojectbackend.exception.ResourceNotFoundException;
 import com.example.ngoprojectbackend.model.Event;
 import com.example.ngoprojectbackend.service.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +13,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+@CrossOrigin(origins = "http://localhost:4200") // replace with update url when pushed to AWS EC2 server
 @RestController
 @Transactional
 public class EventMgmtController {
@@ -26,11 +28,14 @@ public class EventMgmtController {
 
     // get Event by id rest api
     @GetMapping("/getEvent/{id}")
-    public ResponseEntity<Event> getEventById(@PathVariable int id) {
+    public ResponseEntity<Event> getEventById(@PathVariable int id) throws ResourceNotFoundException {
         Event event = eventService.getEventById(id);
-        // after crating exception handler
-        // create custom exception if event with id does not exist
-        return ResponseEntity.ok(event);
+
+        if (id != event.getEventId()) {
+            throw new ResourceNotFoundException("Event id: " + id + "does not exist.");
+        } else {
+            return ResponseEntity.ok(event);
+        }
     }
 
     // create event rest api
@@ -41,38 +46,39 @@ public class EventMgmtController {
 
     // update event rest api
     @PutMapping("/updateEvent/{id}")
-    public ResponseEntity<Event> updateEvent(@PathVariable int id, @RequestBody Event eventInfo) {
+    public ResponseEntity<Event> updateEvent(@PathVariable int id, @RequestBody Event eventInfo) throws ResourceNotFoundException {
         Event event = eventService.getEventById(id);
 
-        // after crating exception handler
-        // create custom exception if event with id does not exist
+        if (id != event.getEventId()) {
+            throw new ResourceNotFoundException("Event id: " + id + "does not exist.");
+        } else {
+            event.setEventName(eventInfo.getEventName());
+            event.setCategory(eventInfo.getCategory());
+            event.setLocation(eventInfo.getLocation());
+            event.setStartDate(eventInfo.getStartDate());
+            event.setEndDate(eventInfo.getEndDate());
+            event.setStartTime(eventInfo.getStartTime());
+            event.setEndTime(eventInfo.getEndTime());
+            event.setAdultTicketPrice(eventInfo.getAdultTicketPrice());
+            event.setChildTicketPrice(eventInfo.getChildTicketPrice());
 
-        event.setEventName(eventInfo.getEventName());
-        event.setCategory(eventInfo.getCategory());
-        event.setLocation(eventInfo.getLocation());
-        event.setStartDate(eventInfo.getStartDate());
-        event.setEndDate(eventInfo.getEndDate());
-        event.setStartTime(eventInfo.getStartTime());
-        event.setEndTime(eventInfo.getEndTime());
-        event.setAdultTicketPrice(eventInfo.getAdultTicketPrice());
-        event.setChildTicketPrice(eventInfo.getChildTicketPrice());
-        event.setImage(eventInfo.getImage());
-
-        Event updatedEvent = eventService.createEvent(event);
-        return ResponseEntity.ok(updatedEvent);
+            Event updatedEvent = eventService.createEvent(event);
+            return ResponseEntity.ok(updatedEvent);
+        }
     }
 
     // delete event rest api
     @DeleteMapping("/deleteEvent/{id}")
-    public ResponseEntity<Map<String, Boolean>> deleteEvent(@PathVariable int id) {
+    public ResponseEntity<Map<String, Boolean>> deleteEvent(@PathVariable int id) throws ResourceNotFoundException {
         Event event = eventService.getEventById(id);
 
-        // after crating exception handler
-        // create custom exception if event with id does not exist
-
-        eventService.deleteEvent(event.getEventId());
-        Map<String, Boolean> response = new HashMap<>();
-        response.put("deleted", Boolean.TRUE);
-        return ResponseEntity.ok(response);
+        if (id != event.getEventId()) {
+            throw new ResourceNotFoundException("Event id: " + id + "does not exist.");
+        } else {
+            eventService.deleteEvent(event.getEventId());
+            Map<String, Boolean> response = new HashMap<>();
+            response.put("deleted", Boolean.TRUE);
+            return ResponseEntity.ok(response);
+        }
     }
 }
